Fix tests to allow SDKs without Qt
The tests were written so that they assumed that the SDK would have
Qt installed in it, but not all SDKs delivered by the PELUX project
do have Qt installed, some of them are headless and don't need Qt.

This patch rewrites the test framework in a way so that one can
easier add new functionality to the SDE and write new tests for
it.

Signed-off-by: Richard Pannek <richard.pannek@pelagicore.com>